Move Groups tests into new structure
