+++ tests/protocols/gmpv208/entities/groups/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2018-2021 Greenbone Networks GmbH
+# Copyright (C) 2021 Greenbone Networks GmbH
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 #
@@ -16,20 +16,9 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from gvm.errors import RequiredArgument
-
-
-class GmpCloneGroupTestCase:
-    def test_clone(self):
-        self.gmp.clone_group('a1')
-
-        self.connection.send.has_been_called_with(
-            '<create_group>' '<copy>a1</copy>' '</create_group>'
-        )
-
-    def test_missing_id(self):
-        with self.assertRaises(RequiredArgument):
-            self.gmp.clone_group('')
-
-        with self.assertRaises(RequiredArgument):
-            self.gmp.clone_group(None)
+from .test_clone_group import GmpCloneGroupTestMixin
+from .test_create_group import GmpCreateGroupTestMixin
+from .test_delete_group import GmpDeleteGroupTestMixin
+from .test_get_group import GmpGetGroupTestMixin
+from .test_get_groups import GmpGetGroupsTestMixin
+from .test_modify_group import GmpModifyGroupTestMixin
