+++ tests/protocols/gmpv208/entities/groups/test_clone_group.py
@@ -16,27 +16,20 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from gvm.errors import GvmError
+from gvm.errors import RequiredArgument
 
 
-class GmpDeleteGroupTestCase:
-    def test_delete(self):
-        self.gmp.delete_group('a1')
-
-        self.connection.send.has_been_called_with(
-            '<delete_group group_id="a1" ultimate="0"/>'
-        )
-
-    def test_delete_ultimate(self):
-        self.gmp.delete_group('a1', ultimate=True)
+class GmpCloneGroupTestMixin:
+    def test_clone(self):
+        self.gmp.clone_group('a1')
 
         self.connection.send.has_been_called_with(
-            '<delete_group group_id="a1" ultimate="1"/>'
+            '<create_group>' '<copy>a1</copy>' '</create_group>'
         )
 
-    def test_missing_group_id(self):
-        with self.assertRaises(GvmError):
-            self.gmp.delete_group(None)
+    def test_missing_id(self):
+        with self.assertRaises(RequiredArgument):
+            self.gmp.clone_group('')
 
-        with self.assertRaises(GvmError):
-            self.gmp.delete_group('')
+        with self.assertRaises(RequiredArgument):
+            self.gmp.clone_group(None)
