+++ tests/protocols/gmpv208/entities/groups/test_delete_group.py
@@ -16,30 +16,27 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+from gvm.errors import GvmError
 
-class GmpGetGroupsTestCase:
-    def test_get_groups(self):
-        self.gmp.get_groups()
 
-        self.connection.send.has_been_called_with('<get_groups/>')
-
-    def test_get_groups_with_filter_string(self):
-        self.gmp.get_groups(filter_string='foo=bar')
+class GmpDeleteGroupTestMixin:
+    def test_delete(self):
+        self.gmp.delete_group('a1')
 
         self.connection.send.has_been_called_with(
-            '<get_groups filter="foo=bar"/>'
+            '<delete_group group_id="a1" ultimate="0"/>'
         )
 
-    def test_get_groups_with_filter_id(self):
-        self.gmp.get_groups(filter_id='f1')
+    def test_delete_ultimate(self):
+        self.gmp.delete_group('a1', ultimate=True)
 
-        self.connection.send.has_been_called_with('<get_groups filt_id="f1"/>')
+        self.connection.send.has_been_called_with(
+            '<delete_group group_id="a1" ultimate="1"/>'
+        )
 
-    def test_get_groups_with_trash(self):
-        self.gmp.get_groups(trash=True)
+    def test_missing_group_id(self):
+        with self.assertRaises(GvmError):
+            self.gmp.delete_group(None)
 
-        self.connection.send.has_been_called_with('<get_groups trash="1"/>')
-
-        self.gmp.get_groups(trash=False)
-
-        self.connection.send.has_been_called_with('<get_groups trash="0"/>')
+        with self.assertRaises(GvmError):
+            self.gmp.delete_group('')
